Handle symbolic links, eliminate spurious reparse.
Only a single level of symbolic links are handled.
Changes either in the linked file or the link itself will trigger
reparsing the file (before this patch, if a symbolic link was used,
changes in the linked files were not detected).

Do a better job at filtering the Gio.file.monitor_file() events
to avoid reparsing the file needlessly.